Went back from APScheduler method
APScheduler won't work well for apps on free dynos.

+++ data_storage.py
@@ -26,24 +26,32 @@ class Media(db.Model):
 media_finder = MediaFinder()
 
 
-def update_all_info():
+def update_all_info(code):
     """Gets all required data from the TMDB API and updates the information in the database."""
-    # Clear all previous data from the data table
-    db.session.query(Media).delete()
+    # If the code supplied is correct, update the database.
+    if code == os.environ.get("UPDATE_VERIFICATION_CODE"):
 
-    # Add the data for the main homepage
-    main_home = Media(page_name="main_home", page_data=get_main_home_info())
-    db.session.add(main_home)
+        # Clear all previous data from the data table
+        db.session.query(Media).delete()
 
-    # Add the data for the movie homepage
-    movie_home = Media(page_name="movie_home", page_data=get_movie_home_info())
-    db.session.add(movie_home)
+        # Add the data for the main homepage
+        main_home = Media(page_name="main_home", page_data=get_main_home_info())
+        db.session.add(main_home)
 
-    # Add the data for the tv homepage
-    tv_home = Media(page_name="tv_home", page_data=get_tv_home_info())
-    db.session.add(tv_home)
+        # Add the data for the movie homepage
+        movie_home = Media(page_name="movie_home", page_data=get_movie_home_info())
+        db.session.add(movie_home)
 
-    db.session.commit()
+        # Add the data for the tv homepage
+        tv_home = Media(page_name="tv_home", page_data=get_tv_home_info())
+        db.session.add(tv_home)
+
+        db.session.commit()
+
+        return True
+
+    else:
+        return False
 
 
 def load_info(page_name):
